backend/requirements.txt
fastapi
uvicorn
python-dotenv
crewai-tools
langchain-litellm
google-generativeai
crewai[google-genai] 
langchain-google-genai
pydantic